System testing of nanoDFT demo notebook
This PR adds an initial workflow that uses the IPU model to check that
the nanoDFT demo notebook doesn't explode.

Some of the changes were necessary to ensure that everything runs with a
configurable number of tiles as well as working around the lack of
executable caching available on the IPU model.

This PR also made some changes to try and support testing the dataset
generation notebook.  This notebook is still failing and will need to be
addressed in a later PR if we want to extend this sort of system level
testing.